+++ src/lib/actions/tasks.ts
@@ -2,6 +2,7 @@
 
 import { PrismaClient } from '@prisma/client';
 import { TaskPayload, payloadToTask } from '../types';
+import { getServerUser, validateServerUser } from '../auth';
 import { revalidatePath } from 'next/cache';
 
 const prisma = new PrismaClient();
@@ -10,11 +11,17 @@ const prisma = new PrismaClient();
  * Write a new task into the database, creating new subtasks when needed
  */
 export async function createTask(task: TaskPayload) {
+    const user = await getServerUser();
+    if (!user) {
+        return;
+    }
+
     // create new task
     const createdTask = await prisma.task.create({
         data: {
             ...payloadToTask(task),
             id: undefined, // generate task id
+            userId: user.id,
         },
     });
 
@@ -38,6 +45,10 @@ export async function createTask(task: TaskPayload) {
  * Handles creating, updating, and deleting any subtasks as needed
  */
 export async function updateTask(task: TaskPayload) {
+    if (!(await validateServerUser())) {
+        return;
+    }
+
     // update the task
     const updatedTask = await prisma.task.update({
         where: { id: task.id },
@@ -80,6 +91,10 @@ export async function updateTask(task: TaskPayload) {
  * Delete a task in the database, given by task id
  */
 export async function deleteTask(id: string) {
+    if (!(await validateServerUser())) {
+        return;
+    }
+
     await prisma.task.delete({ where: { id } });
 
     revalidatePath('/tasks');
@@ -92,6 +107,10 @@ export async function deleteTask(id: string) {
  * setting all children to match the parent task's completion state
  */
 export async function setTaskCompletion(id: string, completed: boolean) {
+    if (!(await validateServerUser())) {
+        return;
+    }
+
     await prisma.task.update({
         where: { id },
         data: { completed },
@@ -114,6 +133,10 @@ export async function setTaskCompletion(id: string, completed: boolean) {
  * - If any sibling subtask is not completed, set parent to not completed
  */
 export async function setSubtaskCompletion(id: string, completed: boolean) {
+    if (!(await validateServerUser())) {
+        return;
+    }
+
     // update just this subtask's completion state
     const subtask = await prisma.subtask.update({
         where: { id },
